Bug 1198458: Webrtc updated to branch 43; pull made 2015-09-29 09:00AM PDT rs=jesup https://chromium.googlesource.com/external/webrtc git checkout -b my_branch refs/remotes/branch-heads/43 commit 6979b024d7cebfdcd1e8f66da59ea157bbc9e47e (I believe)
UltraBlame original commit: f62c9e49a44fb473b5c8701776d8f5ae3da22cbc

+++ media/webrtc/trunk/webrtc/modules/audio_coding/main/acm2/acm_send_test.cc
@@ -61,7 +61,7 @@ bool AcmSendTest::RegisterCodec(int codec_type,
 
 Packet* AcmSendTest::NextPacket() {
   assert(codec_registered_);
-  if (filter_.test(payload_type_)) {
+  if (filter_.test(static_cast<size_t>(payload_type_))) {
     
     
     
@@ -94,7 +94,7 @@ int32_t AcmSendTest::SendData(FrameType frame_type,
                               uint8_t payload_type,
                               uint32_t timestamp,
                               const uint8_t* payload_data,
-                              uint16_t payload_len_bytes,
+                              size_t payload_len_bytes,
                               const RTPFragmentationHeader* fragmentation) {
   
   frame_type_ = frame_type;
@@ -111,7 +111,7 @@ Packet* AcmSendTest::CreatePacket() {
   uint8_t* packet_memory = new uint8_t[allocated_bytes];
   
   packet_memory[0] = 0x80;
-  packet_memory[1] = payload_type_;
+  packet_memory[1] = static_cast<uint8_t>(payload_type_);
   packet_memory[2] = (sequence_number_ >> 8) & 0xFF;
   packet_memory[3] = (sequence_number_) & 0xFF;
   packet_memory[4] = (timestamp_ >> 24) & 0xFF;
